refactor: 移除不必要的类型转换以简化代码
移除 `test_true` 和 `test_pred` 在计算准确率时的冗余 `astype(np.int64)` 转换，使代码更简洁且不影响功能。

+++ main_cls.py
@@ -165,8 +165,8 @@ def train(args, io):
             test_pred.append(preds.detach().cpu().numpy())
         test_true = np.concatenate(test_true)
         test_pred = np.concatenate(test_pred)
-        test_acc = metrics.accuracy_score(test_true.astype(np.int64), test_pred.astype(np.int64))
-        avg_per_class_acc = metrics.balanced_accuracy_score(test_true.astype(np.int64), test_pred.astype(np.int64))
+        test_acc = metrics.accuracy_score(test_true, test_pred)
+        avg_per_class_acc = metrics.balanced_accuracy_score(test_true, test_pred)
         # 记录测试指标
         writer.add_scalar('Test/Loss', test_loss/count, epoch)
         writer.add_scalar('Test/Acc', test_acc, epoch)
@@ -223,8 +223,8 @@ def test(args, io):
         test_pred.append(preds.detach().cpu().numpy())
     test_true = np.concatenate(test_true)
     test_pred = np.concatenate(test_pred)
-    test_acc = metrics.accuracy_score(test_true.astype(np.int64), test_pred.astype(np.int64))
-    avg_per_class_acc = metrics.balanced_accuracy_score(test_true.astype(np.int64), test_pred.astype(np.int64))
+    test_acc = metrics.accuracy_score(test_true, test_pred)
+    avg_per_class_acc = metrics.balanced_accuracy_score(test_true, test_pred)
     outstr = 'Test :: test acc: %.6f, test avg acc: %.6f'%(test_acc, avg_per_class_acc)
     io.cprint(outstr)
